Keep agent tools bound to active sessions
