+++ packages/server/src/server/agent/providers/opencode-agent.ts
@@ -78,6 +78,11 @@ import {
   type OpenCodeRuntime,
   type OpenCodeServerAcquisition,
 } from "./opencode/runtime.js";
+import {
+  bindOpenCodePaseoToolingSession,
+  ensureOpenCodePaseoToolingRuntime,
+  withOpenCodePaseoToolingEnv,
+} from "./opencode/paseo-tooling.js";
 import { normalizeProviderReplayTimestamp } from "../provider-history-timestamps.js";
 import { revertOpenCodeConversationAndFiles } from "./opencode/rewind.js";
 
@@ -1280,9 +1285,14 @@ export class OpenCodeAgentClient implements AgentClient {
     options?: AgentCreateSessionOptions,
   ): Promise<AgentSession> {
     const openCodeConfig = this.assertConfig(config);
+    const paseoToolingRuntime = ensureOpenCodePaseoToolingRuntime();
     const acquisition = await this.runtime.acquireServer({
       force: false,
-      env: launchContext?.env,
+      env: withOpenCodePaseoToolingEnv({
+        env: launchContext?.env,
+        launchContext,
+        runtime: paseoToolingRuntime,
+      }),
     });
     const { url } = acquisition.server;
     const client = this.runtime.createClient({
@@ -1307,6 +1317,10 @@ export class OpenCodeAgentClient implements AgentClient {
       }
 
       await this.populateModelContextWindowCache(client, openCodeConfig.cwd);
+      const unbindPaseoToolingSession = bindOpenCodePaseoToolingSession({
+        launchContext,
+        sessionId: session.id,
+      });
 
       return new OpenCodeAgentSession(
         openCodeConfig,
@@ -1317,6 +1331,7 @@ export class OpenCodeAgentClient implements AgentClient {
         acquisition.release,
         options?.persistSession,
         launchContext?.agentId,
+        unbindPaseoToolingSession,
       );
     } catch (error) {
       acquisition.release();
@@ -1342,7 +1357,15 @@ export class OpenCodeAgentClient implements AgentClient {
       cwd,
     };
     const openCodeConfig = this.assertConfig(config);
-    const acquisition = await this.runtime.acquireServer({ force: false });
+    const paseoToolingRuntime = ensureOpenCodePaseoToolingRuntime();
+    const acquisition = await this.runtime.acquireServer({
+      force: false,
+      env: withOpenCodePaseoToolingEnv({
+        env: launchContext?.env,
+        launchContext,
+        runtime: paseoToolingRuntime,
+      }),
+    });
     const { url } = acquisition.server;
     const client = this.runtime.createClient({
       baseUrl: url,
@@ -1351,6 +1374,10 @@ export class OpenCodeAgentClient implements AgentClient {
 
     try {
       await this.populateModelContextWindowCache(client, openCodeConfig.cwd);
+      const unbindPaseoToolingSession = bindOpenCodePaseoToolingSession({
+        launchContext,
+        sessionId: handle.sessionId,
+      });
 
       return new OpenCodeAgentSession(
         openCodeConfig,
@@ -1361,6 +1388,7 @@ export class OpenCodeAgentClient implements AgentClient {
         acquisition.release,
         undefined,
         launchContext?.agentId,
+        unbindPaseoToolingSession,
       );
     } catch (error) {
       acquisition.release();
@@ -2802,6 +2830,7 @@ class OpenCodeAgentSession implements AgentSession {
     releaseServer?: () => void,
     persistSession = true,
     private readonly agentId?: string,
+    private readonly unbindPaseoToolingSession?: () => void,
   ) {
     this.config = config;
     this.client = client;
@@ -3574,6 +3603,7 @@ class OpenCodeAgentSession implements AgentSession {
       await this.deleteProviderSessionIfEphemeral();
       this.activeForegroundTurnId = null;
     } finally {
+      this.unbindPaseoToolingSession?.();
       this.releaseServer?.();
       this.releaseServer = null;
     }
